pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "asyncpg>=0.30.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "pytz>=2025.2",
    "scipy>=1.16.2",
    "uvloop>=0.22.1",
    "websockets>=15.0.1",
]
